Determ Runs Update - Bug Fix in get.schoolvacc.time.age.specific()
The previous commit was code used to generate preliminary report for the 2023 Q4 report in January 2024.  Since then I found a bug in the code, specifically in get.schoolvacc.time.age.specific() where roughly only 1/12 of probability of school entry vaccination was being applied.  I code was updated.  This set of code was used to generate update preliminary report shared with the Steering Committee 8 April 2024.